Remove need for plugins to match core version
This change was possible by making BREAKING changes to the core again. We reverted the requirement to specify a different main function in the metadata and went back to using _prolog as the entrypoint. Instead, we now pass an API object as a parameter that is used to do all the hooking. Thus bypassing the need to link against sy_core directly and making it possible to not require matching versions